update: global telegram refactoring + localization added (but not everywhere)
