+++ src/telegram_bot_service/services/game_service.py
@@ -1,11 +1,9 @@
 from typing import List, Dict
 from database.database import games
-from services.utils import async_wait
 
 
 class GameService:
     @classmethod
-    @async_wait()
     async def get_games(cls, filters: Dict):
         print(filters, flush=True)
         offset = filters.get("offset", 0)
@@ -28,12 +26,10 @@ class GameService:
         )
 
     @classmethod
-    @async_wait()
     async def get_game_by_id(cls, id: str) -> Dict:
         return list(filter(lambda x: x["id"] == id, games))[0]
 
     @classmethod
-    @async_wait()
     async def get_games_by_str_in_title(cls, title: str) -> List[Dict]:
         title = title.strip().lower()
         return list(filter(lambda x: title in x["title"].lower(), games))
